Fix clist closepath to wrong start point. Bug 695205, cust 870.
When writing paths into the clist, we simplify the path for each band.
This means that whole line segments that are not present within a band
can be omitted which speeds up path rendering for bands. This code had
a bug that could cause the closepath of a subpath to connect to the
wrong point.

If the last element written for a subpath ends at point X, and the
start point of the first element of the next subpath is that same
point X, then the code would forget to emit a "catchup" moveto.
This would cause a closepath to connect to the initial point of
the previous subpath.

Amazingly, this bug has been there since rev 3.60 (Nov 1995) although
a partial fix was added in 3.68 (May 1996).