PDF interpreter - work around GSView usage of the interpreter
Bug #695179 "Inert links in PDF and collapsing PDF when opened in gsview with gs ver. newer that 9.07"

Commit 073f460af5bb37edb1849c5d6235048598100437 introduced a cumulative page
count, which is added to pdfmarks when encountered in order to cope with
running multiple input files. The pdfmarks from the second file would
be incorrect unless the number of pages in the first file were added to the
destinationpage numbers.

Unfortunately the (undocumented, unsupported) way that GSView works with the
PDF interpreter to retrieve links failed, because the variable is unavailable
under the conditions that GSView exercises the interpreter.

In this commit we extend the CumultaivePageCount usage so that we first check
if the variable is defined, and if its not, we set a default of '0'. This
should mean that the supported use of the PDF interpreter continues to work
correctly with multiple input files, and restores the GSView behaviour.

No differences expected.