CUPS Raster output device: Fix cupsRasterOpen() call to get correct PWG Raster output
The CUPS Raster output device generated incorrect PWG Raster output
because it called cupsRasterOpen() without the required
CUPS_RASTER_WRITE_PWG mode setting. Due to the output never being
tested with an IPP Everywhere printer but only with software which
also accepts CUPS Raster, the bug did not get discovered until testing
the output against the mime type recognition of CUPS where it failed.

Note that this patch only fixes the problem if a shared libcupsimage
library of the system is used, not with the library included with
Ghostscript.  This is because the included library is too old to
contain PWG Raster output support. If the included CUPS libraries have
to be used, the CUPS Raster output device is built as before to assure
that Ghostscript builds and CUPS Raster can still get generated.