pdfwrite - add an idiom for certain patterns produced by the MS driver
Bug #695230 "Invalid pattern fill rendering"

The Windows PostScript driver produces some patterns (for background fills)
which vary the size of the pattern cell according to the device resolution.
This means that as the resolution goes up, more pattern cells are drawn,
resulting in a finer and finer pattern.

This resolution neatly defeats our high level pattern capture in pdfwrite.
It only works properly if the device is running at 72 dpi, and the example
runs at 600.

It seems the easiest way to deal with this is to create an idiom which will
replace the resolution-varying pattern with a fixed pattern (it isn't possible
to have a resolution-dependent pattern in PDF anyway) and this commit adds
such an idiom.

It also updates the idioms to use the new .spec_op operator to interrogate
device parameters rather than the slow getdeviceparams previously used.

No differences expected.