pdfwrite - better handling of input PDF files with broken fonts
Bug #695216 "Spurious errors and bad PDF written from input file comparefles/Bug693711.pdf"

The errors are not spurious, and are caused by the missing/damaged fonts in
the input file.

This commit chooses to ignore these errors, but I may rethink that if it causes
future problems, the fact that there are no errors when rendering does not
mean that there should be no errors when creating a PDF file.

Because the PDF interpreter ignores errors and continues regardles we were
then attempting to write an invalid font, which failed, resulting in an
(unused) entry in the xref with a duplicate offset of an earlier entry. This
commit further refines the broken font handling by detecting the fact that
the font was incorrectly written and elides it from the xref table.

This addresses both of the concerns raised in the bug report, though I would
point out that ignoring errors when the output device is pdfwrite is always
very likely to produce broken PDF files.

This does alter the font substitution for Bug693711.pdf on Linux.