pdfwrite - improved handling of 'glyphshow' and similar operations
Bug #695168 "Problem converting xps to pdf"
Bug #695259 "both (incorrect) B/W and (correct) AA rendering of Libertine font in same PDF output"

Although these are different products, one being Ghostscript and the other
gxps, the underlying problem is similar. The PostScript makes extensive use
of the glyphshow operator, which ends up as a TEXT_FROM_SINGLE_GLYPH operation
in the graphics library, and the XPS interpreter always uses TEXT_FROM_GLYPHS.

In both cases the font is effectively unencoded when pdfwrite seee it. Since
we cannot construct and use an unencoded font in PDF we have no alternative
but to create an encoding for the font, and write the text using that encoding.

The way this is done we use the name table, which effectively means that the
character code we use is derived from the first byte of the glyph name. For
limited usage this works well, but more complex usage can result in problems.
For example the glyphs /o and /omicron are both encoded at index 103.

Previously this would cause us to fall back to rendering the font and embedding
as a type 3 font, or in the worst case an inline image. This is because when
encoding the text we would discover that an the font already had a glyph
encoded at the correct index, and would simply pass on, not realising it was
the incorrect glyph. Later we would check the actual glyph index against the glyph index
of the glyph encoded at that position, realise they were different and throw an error.

In this commit we check the glyph index early, at the point where we encode the text.
If the font already has a glyph encoded at the given character code, we check
the glyph  index to see if it matches the current glyph. If it does all is well
but if it doesn't we break out and create a new font instance, with the new
glyph encoded in it.

Potentially this could result in a *lot* of font subsets being created, which
would increase the size of the output PDF file, but the quality improvement
is well worth it.

No differences expected.