PDF interpreter - pass image height as a /DecodeParm to all filters
Bug #695116 "**** Warning: File has insufficient data for an image"

This is not a solution for this bug, merely framework to allow it to be fixed.

The PDF file contains a DCT encoded stream for an image, where the header
declares a height of '-1' and then later uses a DNL marker to set the actual
number of rows of data in the image. This does not work with libjpeg (in
fact it doesn't work with an number of JPEG consumers).

However, because this is an image in a PDF file, we do have the image dimensiosn
from the image dictionary, so we can use that instead, if only we had a way
to pass the information into the stream decode filter.

This commit modifies the dictionary we provide to the stream code in the PDF
interpreter, we now insist that all the Filters in an image shall have a
DecodeParms dictionary, if they don;t already have one we create them as
required. In that dictionary we add a /Height key and put the Height retrieved
from the image dictionary in it.

All the decode operators handle being passed a dictionary with parameters,
so all that is then necessary is to have the DCT decode filter check the
parameter dictionary and store the Height from it in a new entry in the DCT
stream state.

The parameter is now available for use by the DCT decode filter, and a future
commit will use it to fix the actual problem.

No differences expected.