Bug 695116: Rewrite JPEG header height when required.
This follows on from Kens previous commit and should fix the bug.

When a PDF includes a JPEG it states the height to use for that
JPEG. We now check the JPEG header to see if the height declared
there is 0 or is greater than the PDF specified height. If so,
we rewrite the header to contain the height specified in the PDF.

Short of performing radical surgery on libjpeg to make it honour
the DNL tag, this is the best we are going to be able to do.