PDF interpreter - ignore broken appearance dicts
Bug #692181 " %%EOF marker"

The bug title is very misleading, there were two problems;
1) Ghostscript didn't synthesise appearances fro FreeText annotations
2) The appearance dictionaries were invalid.

Previous work added support for FreeText annotations, this commit adds some
slight improvements to that, and in addition it does minimal validation on
the AP (appearance) dictionary. If its invalid we don't bother to use it
but synthesise an appearance anyway.