PDF interpreter - improve synthesis of text form fields with no /AP
Bug #694738 " Text synthesis of missing appearances in the PDF interpreter does not handle UTF16BE"

Previously we didn't strip the 0xFE 0xFF byte order mark from UTF16-BE strings
this commit removes them. We also remove the requirement for a /DA (Default
Appearance) to be present in a Text field before we synthesise a replacement
for a missing AP, taking int account any UTF16 strings.

This does not 'fix' bug #694738 but this is because the /DA uses a font which
is not correctly encoded to draw UTF16 text. The font uses an /Identity-H
CMap which is unlikely to work with UTF16. Replacing it with our own fallback
CID font and using a /Identity-UTF16-H CMap draws the text correctly, which
shows that the problem is the font. NB Acrobat (as usual) ignores appearances
and creates its own representation, so the Acrobat display is not an indication
of a bug in GS.

A few files in the test suite show differences because we no longer attempt
to print the BOM.