pdfwrite - improved handling of images with > 8 BPC
Bug #691713 "/rangecheck in fts_18_1805.pdf on pdfwrite device"
Bug #694731 "pdfwrite doesn't handle 16 bpc images"

To date images with more than 8 bits per component, using any transparency
features, resulted in a rangecheck error as the code was incapable of
handling 12 or 16 bpc.

The new image code (improved for colour management) now allows us to apply
a simple 16->8 (discarding the lower 8 bits) subsampler and this results
in all the files which caused problems working correctly.

In addition, many files which previously 'fell back' to the default implementation
(where each image sample is transformed into a filled rectangle) now preserve
the image as an image. This makes the resulting PDF or PostScript file
considerably smaller and quicker to process.

Note that this *only* works if the ColorConversionStrategy is 'LeaveColorUnchanged'
This is because the existing image colour space conversion code cannot cope
with anything other than 8-bit input.

A number of test files show differences.

The majority are PDF files which previously had content missing but now
render correctly.

A number are due to the fact that the image is no longer degenerated into
thousands of rectangles which means that it is recompressed, if the input
was JPEG and the compression is also JPEG this results in visible artefacts.

Some files show differences due to the fact that the naive downsampling
gives slightly different results from the accurate colour conversion. These
show slight improvements in about as many cases as they show slight degradation