pdfwrite - don't include 'GoToR' references in check of destinations
Bug #695269 " PDF file containing outline or link annotation with GoToR action processing ended with a message "ERROR: A pdfmark destination page XXX points beyond the last page YYYY.""

The problem is that the function which gets a numeric page number, either by
finding the /Next or /Prev keys or by finding an absolute page number, was
also updating the 'max_referred_page' variable. This is checked when closing
the PDF file and an error message emitted if it exceeds the number of pages
in the PDF file.

Of course, for a GoToR action, its valid for the destination (which is in a
different file after all) to be greater than the number of pages in the current
file.

This commit removes the update for the max_referred_page variable from
pdfmark_page_number, and places it in the calling function. This makes more
sense to me, its not reasonable that s utility function should have an
unexpected side effect like that, and it makes more sense for the function
which handles the destination to decide whether it should update this variable.

No differences expected, we don't test this on the cluster.