Check for trivially clipped out images before checking trivially non-clipped
Bug #695320 "Difference in bbox result from 9.06 onwards"

commit a13600f1c241c3e36dbe4973e9d78a8934b16004 introduced checks to see
if we actually needed a clip device before going ahead and installing it.
However, the order of checks was slightly incorrect. The code clamps the
rectangle to be tested to the outer box of the clip path, then checked to
see if the rectangle was completely included in the inner box.

If the object was completely clipped out then the rectangle would be clamped
to the outer box of the clip. If the outer and inner boxes were the same
(rectangular clip) then the test could decide that the totally clipped out
rectangle was in fact completely enclosed in the inner box.

We fix this here by checking to see if the image is clipped out (by testing
to see if the top/bottom or left/right edges are co-incident) before we
check for trivial inclusion.

For totally clipped out objects we then abort the object which, like the
original fix prevents us installing a clip device and also prevents us
going through the trapezoid degeneration., Potentially a small performance
win, as well as getting the bbox correct.

No differences expected