pdfwrite - preserve enumerator text for composite data
Bug #695322 "Distilling PS to PDF using GS causes strange characters for spaces when using a particular sequence in the PS"

The problem occurs because when processing text in a composite font pdfwrite
overwrites the string data while working on it. In general this is not actually
a problem, as the string data argument to the show family of operators is
normally not reused.

However, in this case the customer's PostScript inserts space glyphs by calling
a bound procedure, instead of simply calling show. The first execution works
properly, but corrupts the string, the second execution uses the corrupted
string as the data, and is unable to find a glyph in the font to match the
corrupted data, which results in the /.notdef being used instead.

The particular routine in question (scan_cmap_text) is one of the most
convoluted and bloated routines in pdfwrite, and after some inspection it
was obviously going to be very difficult to avoid overwriting the string
data, it would run considerable risk of breaking the code in subtle ways,
and would in any event mean copying the string data around, possibly mutliple
times.

Accordingly I've chose to make a new copy of the string data in
process_cmap_text, which is the only parent of scan_cmap_text, and save/restore
the string data round the call. This prevents the bound string being
corrupted and fixes the problem.

This also exhibits a progression in Bug691680.ps and I've noted in that
thread that the remaining problem is fixed.