pdfwrite colour conversion - handle alternate conversion to sRGB from DeviceN
Bug #695316 "A DeviceN image in pdf can't be converted to RGB image, it is converted to many paths instead"
Bug #695306 "Converting a PDF to RGB colorspace causes missing drawing parts"

The code for converting a DeviceN alternate space and tint transform wasn't
catering for a ColorConversionStrategy of /sRGB. This commit ads the
ccs_sRGB case, and converts it to RGB as per the documentation, using the
selected RGB profile.

Bug 695316 now converts the image to an image, though it produces an error
on exit relating to ICC profiles.

Bug 695306, for me, caused a seg fault, which this commit corrects. I do not
see any missing content with either sRGB or RGB conversion. I do see the
same error as for Bug #695316 and I'll continue working on that problem
under that bug number

No differences expected, the c;luster doesn't test colour conversion