Add pageneutralcolor as a printer device parameter to allow query.
fillpage (erasepage from PS) will re-enable monitoring so we needn't
ouptut the page in order to restart monitoring.

This allows PostScript or clients to use get_params or get_param to
read the pageneutralcolor state. Presumably this would be used prior
to showpage. Coupled with the EndPage proc, PS can examine this even
without emitting the pages.

Note that this is only accurate when in clist mode, so -dMaxBitmap=0
and -dGrayDetection=true should both be used.