Apply cat02 chromatic adaptation to primaries when creating ICC profile from CalRGB
While I had pickled the CAM into the transform created using the lutAtoB type, I
failed to apply the transform for the matrix based solution that is used for the
profile when created from CalRGB.  I did have a comment in the code that this needed
to be done.

Several progressions occur.