Default PCL to full bleed, setting the printable region margins to 0.
Actually this change was partially implemented previously, but there
were still lingering parts of the code that used the 1/6" border to
affect the output.  Older HP printers default to a 1/6" inch printable
region.  This can be implemented by changing the default HWMargins in
the device or by setting them with the command line option -H, see the
PCL documentation for details.