pdfwrite - colour conversion of DeviceN alternate spaces
Bug #695306 "Converting a PDF to RGB colorspace causes missing drawing parts"

There are two problems addressed in this commit.

Firstly the code wasn't catering properly for the possibility of an /Indexed
DeviceN space. We did dereference the Indeexed space, but not until after
we had tried to use the original space as a DeviceN, with unfortunate
potential results. We move the dereference to the top of the function
convert_DeviceN_alternate so that we properly use the DeviceN space.

Secondly when creating the new colour space (ie with a different alternate
space), we add the names of the inks from the original DeviceN space.
However, we were using add_no_copy which meant when the new array was
garbage collected we freed the names, but left them being pointed at by
the original array, leading to pointers to garbage memory.

No differences expected as the colour conversion is not tested by the cluster