pdfwrite - ensure we have V2 ICC profiles for PDF/A-1 and PDF < 1.5
Bug 695118 "PDF/A-file is generated with ICC version 4.2 profiles"

PDF/A-1 requires us to use ICC v2 profiles due to the level of PDF mandated
by the spec. In addition, we should not use V4 profiles when producing
PDF less than 1.5.

Michael added considerable code to allow us to get a V2 profile, given a
V4 profile, and this commit makes the changes required for pdfwrite. Its more
extensive than appears necessary because we need access to the current
imager state in order to generate the V2 profile, so a large number
of routines have been altered to pass along the imager state (or NULL in
some cases, where we can only be dealing with device spaces anyway).

We also change the default version of PDF we produce to 1.5 so that we can
emit V4 profiles and alter the 'prepress' predfined settings similarly.

This causes changes in a Genoa CET file, but its just different, not wrong.