pdfwrite - detect corrupted type 3 fonts and generate an errror
Bug #694186 "Seg faults found by fuzzing in names_index_ref"

The file has a corrupted glyph in a type 3 font, which throws an error.
However, pdfwrite can't tell that the glyph threw an error rather than
simply completing successfully, and tries to use the returned glyph ID. The
glyph ID is GS_NO_GLYPH which fails with a seg fault when looking up the
name.

In this commit we check the glyph ID to see if its GS_NO_GLYPH and throw an
error if it is. This prevents us looking up an invalid glyph ID in the
name table, and the file is demonstrably broken.

This fixes a1c8db... the other file, cc5c... is also fixed, but in newer
code never even makes it to this point as the current code detects a different
problem before it reaches this point.