pdfwrite - fix memory corruption caused by type 3 capture of PCL fonts
Bug #695350 "GhostPCL line shifting"

It was impossible to reproduce this bug exactly as the reporter is running
Ghostscript on an unusual platform. Fortunately what appears to be the same
problem could be reproduced on a PowerPC Mac running Linux. It is possible
that this commit doesn't actually fix the bug though.

When capturing glyph descriptions in order to build a type 3 font from a
PCL font, we were saving the default text procedures from the enumerator
(which was created by gx_default_text_begin), and then restoring them when
we had finished processing the glyph. However, by that time we had already
releases the enumerator, which meant we were corrupting memory.

Desk checking the code, I can't see any reason for saving the procedures
from the text enumerator, and I conclude this is either a holdover from an
earlier version of this code, or an oversight when writing the code that
captured PCL fonts. In any event this commit removes the save/restore
behaviour, since it is definitely causing memory corruption.

No differences expected.