pdfwrite - prevent closing the device if its not open
Bug #694380 "Seg fault in pdfwrite"

The file in question throws an error in pdfwrite when we try to write a
FontDescriptor (the font is broken). The PDF interpreter, however,
often ignores errors and continues trying to process the input file.

When producing one output file this is not a problem, as the error occurs
in pdf_close() and so it all works. However, when producing one file per
page this fails, because we close the device, notice the error, and
do not open it anew. However the PDF interpreter just carries on. When we
come to the end of the next page we try to close a device which is already
closed, and this means we try to dereference an invalid pointer.

In this commit we check the 'is_open' flag of the device in pdf_close and
if the device is not open then we jump straight back out. We also set is_open
to false in the same routine as it wasn't previously being set (because we
couldn't previously close and reopen the device).

This fixes the problem with Bug693711.pdf for me.