Bug 694205: prevent crash rendering unfeasibly large glyph
Initialise the "prev" entry in the top node in the x line list to null and
check it isn't null before we use it. If we reach that point, it indicates that
a) the current path reaches (and probably passes) the limits of our coordinate
space and b) the line list is already in the correct order - as best it can be,
anyway.

No cluster differences.