pdfwrite and ps2write - Mimic the -dNOCACHE output of pswrite
This feature is in response to the thread on gs-devel "pswrite and NOPLATFONTS"

Normally pdfwrite and ps2write try very hard to maintain fonts from the input
as fonts in the output, for reasons of output size, performance and qualiity.

It seems that some people are relying on using pswrite with -dNOCACHE in
order to produce PostScript output with no fonts at all. (As far as I can
see this is to address the limitations of other software). This was never
a goal for pswrite, but it seems that it works 'well enough' to be useful.

Sadly, despite us flagging the fact that pswrite had been deprecated, nobody
gave us any feedback until after (several releases after, in fact) we had
removed the device. We are not going to reinstate support for the pswrite
device, having spent effort to remove it in order to drop support for it,
so this commit adds a new feature which causes pdfwrite and friends to
write the glyph description into the content stream as it is used instead
of constructing fonts.

This is controlled with the new switch NoOutputFonts which defaults to false

There are no cluster differences with this commit. Using the switch causes
pretty much every test file which contains text to render differently.