Avoid failure to clean up temp files when the execution is killed
This currently works only for Windows and Linux builds. For Windows, temp
files are created with the "delete on close" option. For Linux, temp
files are immediately unlinked on creation. In both cases, the initially
opened FILE pointer is kept open throughout execution, and reader access
is provided by duplicating the initial FILE rather than opening a new
FILE pointer based on the file name.