Bug 695363: handle TTF with no/empty loca table.
A broken TT font could end up with us allocating a zero length array for
the for the glyph lengths, then later to attempt to access entries in that
array.

Thus, avoid allocating the zero length array, and treat a NULL glyph_len array
the same as a glyph_len of 0.

No cluster differences.