Bug 695318: mark memory chunks used for single allocations.
Certain chunks in the memory manager are used to contain only one (o_alone)
allocation: "large" allocations and "immovable" allocations specifically.

We need to be able to identify those chunks so we don't end up trying to use
any empty space in them for later allocations - these chhunks can end up with
empty space in them if the original object allocated in them is shrunk.

The problem that occurs is when we free an o_alone object, we immediately free
the chunk containing it, thus any other allocations in that chunk are also
freed.

Adding a c_alone flag to the chunk means we can trivially identify and skip
those chunks when searching for free memory.

No cluster differences.