Bug 695319: Set pdf14 (trans) ctx to NULL on free on error
In the event of an interpreter triggered error during a transparency operation,
pdf14_discard_trans_layer() is called to clean up. It frees the context, but
didn't set the ctx pointer to NULL - thus we could try to free the already
freed context object when the pdf14 compositor device is shut down.

No cluster differences.