Bug 695397: handle "unknown" glyph names >16 chars long
When creating glyph aliases for a TTF, we check for "formatted" unicode glyph
names of the form "uniWXYZ" where "WXYZ" is the glyph code point in base 16.

In doing that we use a temporary string which is 16 characters long (since the
formatted names we're interested won't be longer that 16).

So, check the glyph name is <16 chars long before checking if it's a "uni..."
style name.

No cluster differences.