PDF interpreter - fix Hybrid file parsing
Bug #695403 "Error ocurred while reading XREF table"

The fix for bug #694342 treated an xref section with no entries as an error
and caused the xref to be rebuilt. However, if a PDF file is converted into
a hybrid file (one with an xref *and* an XrefStrm) by adobe CS 3, then it
does this by adding a 'dummy' xref which has a /Prev pointing to the real
xref, and a /XrefStrm pointing to the stream.

The dummy xref has a section "0 0", is first object 0, no objects defined
which falls foul of the test for Bug 694342 and causes the xref to be rebuilt.

This commit adds a boolean which is initially false and is set true in the
parsing loop for the xref section. When we encounter a /trailer, instead
of checking if 0 objects are defined, we now look at that boolean. If its
'true' then we have seen a section entry, and can treat this as valid, even
if no objects are yet defined. If we haven't seen a section then we treat
it as invalid and force an xref rebuild.

This causes sumatra/694313_-_empty_xref.pdf to stop producing errors as
well.