Bug 695394: fix TTF glyph encoding generation
There was a logical error in the glyph encoding generation for TT fonts,
affecting fonts without post tables.

When handling the case where the same glyph name references two different
glyphs, the index in the glyph encoding we wrote the generated name to was the
GID (i.e. the index returned from the cmap table) rather the original index
we were handling from the glyph encoding.

No cluster differences.