ps2write - fix conversion of TrueType fonts with format 0 LOCA tables
Bug #695417 "gs(ps2write) produce invalid ps from certain LibreOffice print emission"

The code in the opdfread prolog for converting TrueType fonts to type 42
fonts was completely broken for fonts with a format 0 LOCA table (16 bit
offsets). The offending routine is StringToLoca.

The code was reading 32 bits instead of 16, which causes it to exhaust
the string data, leaving array members uninitialised, which is the inital
cause of the error. However, it also didn't multiply the offsets it did
read by 2 (format 0 LOCA offsets are multiplied by 2), and had a 'dup'
which resulted in the stack having the wrong value on top on exit.

No differences expected