Revert commit 4fc4b8410d35b193d549558ab212cc727db9460b
See bug #694160 and #695420. Somehow this patch ends up leaving 'quit'
undefined at the end of job, which breaks the '-o' switch. It also somehow
manages to break at least some pdfwrite incantations, leading to an undefined
error with a garbage caller. It kind of looks like the name table is
broken or similar.

Note that the undefined error on quit occurs when running GS interactively
and even when this happens, "systemdict /quit get exec" works, so its not
as simple as /quit being undefined.

No differences expected.