Bug 694160: Prevent malicious code being injected into the system.
This is essentially the same patch as committed in:
4fc4b8410d35b193d549558ab212cc727db9460b

but that patch was missing white space at the ends of automatically
concatenated strings - hence the interpreter saw, for example:
".systemexecend" rather than ".systemexec end".

Additionally, to avoid that confusion in the future, I've removed the reliance
on the automatic string concatenation.

Also removed a spurious additional call to ".uninstallpagedevice".

No cluster differences.