PDF interpreter - read BPC from the JPX stream instead of the resource dict
Bug #695387 "PDF with jpeg2000 conversion problems"

The PDF spec states (table 4.39, BPX on page 340 of the PDF Rerference)
that the Bits Per Component value is to be taken from the JPX stream and
the image dictionary value should be ignored if present. We were only
reading the BPC if either that to the ColorSpace was missing from the
image dictionary.

In this commit we *always* pare the data out of the JPX stream (which is
a small performance penalty). However, this always overwrites the data in
the resource dictionary, which is fine for BPX< but not acceptable for the
colour space, because Acrobat favours the ColorSpace in the
image dictionary over the JPX colour space, and because we set the colour
space to be DeviceGray for softmasks.

So we save the colour space from the resource dictionary (if present), read
the BPC and colour space then, if we had a colour space in the resource
dictionary we restore it back again.

For some reason this causes problems on the cluster with the test file
Bug688845.eps. Since this change only affects JPX images, and PDF files, I
can't see any way that the change can be responsible for this. I am unable
to reproduce the problem on Windows or Linux locally.