pdfwrite - fix a ShadingType 5 with array DataSource and a Function
Bug #695359 "Wrong gradients after conversion to PDF"

The example file is a hand-coded shading of type 5 with the DataSource
being an array, but using a function to convert the parametric value at
each vertex into a colour value.

The PDF specification doesn't support this kind of shading, so we need
to convert it. In doing so we write out a Decode array, which needs to contain
the minimum and maximum colour values. Ordinarily we get these from the
shading dictionary, but when using an array for DataSource this is not
present. In the absence of the Decode array we were defaulting to 0 and 1
but the function actually has a range from 0 to 10. This meant we were
truncating the colour space.

This commit examines the Function (if present) and uses the Domain of the
function as the values instead of 0 and 1.

This may not be sufficient in the absolute general case, but all mesh
shadings are rare in real world files, and mesh shadings using array
DataSource are still less common, as are examples with a function which
does not have a Domain 0-1. So this will suffice until we see more
problems.

Note that Acorbat Distiller X produces different output if you run the
two shadings in the example file separately instead of consecutively. Also
Acrobat X displays the resulting file from both Distiller and Ghostscript
incorrectly (Acrobat XI displays them correctly)

No differences expected.