Bug 695420: Bad pointer to OutputICCProfile string.
This was caused by the gs_setpd.ps logic having a dictionary that has the
OutputICCProfile pointing to dev->icc_struct->device_profile[0]->name, but
when ColorConversionStrategy distillerparam is set, it does a rc_decrement
on the pdev->icc_struct which frees the entire structure, among those the
'name' (which was allocated in non_gc_memory).

Add the OutputICCProfile to the list of .dynamicppkeys in gs_setpd.ps so
that the string will always be updated and the dictionary won't have a
stale pointer.

Having a dict with a string in non_gc_memory is OK I guess, even though
we generally "hide" all non_gc_memory elements from the GC.