pdfwrite - fix handling of Tx fields in AcroForms
Bug #695438 "AcroForm text fields not being rendered"

There are multiple parts to this fix.

1) Alter the default of the NeedAppearances key in AcroForms and AcroForm
field dictionaries to true so that if its not present we will generate
Appearances for fields which have none. (we still won't generate them if
the field is present and set to false, though Acrobat does)

2) Fix the Length key of the Form dictionary we generate for Tx fields,
previously it was set to 0, so the form wasn't ever executed.

3) If there is no 'V' key in the Field dictionary for a Tx field, use the
/DV key instead. If still no key then use an empty string.

These changes uncovered some interesting additional problems.

1) Bug687498 has an AcroForm with a Tx field, declared as multiline, where
the actual data is an empty string. Now that we actually execute the Form
this causes an error. Add a check to treat strings with less than 2 bytes
as single line, so that splitstring doesn't throw an error.

2) Bug692477.pdf, now that we default 'NeedAppearances' to true, we discover
that this file has a Fields array where every element is the null object,
which causes errors trying to extract dictionary key/value pairs from it.
Add a check to see that the Field element is a dictionary, and ignore it if
it isn't.

No differences expected, we don't test with -dShowAcroForm