Fix setpagedevice so that changing ProcessColorModel works
Bug #695439 "Reference to free memory, or segfault with release build"

The problem is the horrendous way our implementation of setpagedevice works.
Where we should query the device for its parameters, our implementation
maintains a dictionary (initially built by querying the device) which is
updated by calls to setpagedevice.

Clearly at some time in the past someone has discovered that this breaks
badly if the device changes a parameter itself, as this won't be reflected
in the saved dictionary. So the '.dynamicppkeys' dictionary was defined,
whose sole purpose is to hold a list of keys whose values must be refreshed
from the device.

However, this dictionary is *also* used to prevent those same keys being
sent to the device.....

This means that commit 2eda51b48c7e97adf7c46fc3d7eb34fcdf6061e6 prevented
any changes to OutputICCProfile being sent to the device.

So I've now made 2 dictionaries with separate purposes, one is for 'read only'
keys which should not be sent to the device, and one which contains the list
of keys which must be read back from the device to replace the stored values
in our dictionary. Keys may appear in both, either, or neither.

The problem in this bug is that we create our initial dictionary when the
device is in DeviceRGB, and we store the RedValues, GreenValues, BlueValues
and GrayValues from the device in our stored dictionary. Then we execute
code to change the ProcessColorModel to DeviceGray. This alters the values
of the RedValues etc in the device. But these are *not* copied into our
stored dictionary. Later we call .installpagedevice which does not change
the ProcessColorModel, but does attempt to set the RedValues etc, which are
invalid for a DeviceGray ProcessCOlorModel, leading to a rangecheck.

I had hoped that adding these keys to the volatile dictionary would be
sufficient to fix the problem, but it seems that we don't update our stored
dictionary with these parameters. The only way to fix this turned out to be
to add the keys to the 'read only keys' dictionary so that we don't try and
send them to the device.

I'm not completely sure this is sensible, it isn't obvious to me what these
values are for, so I can't be certain that they should become 'read only'.
However, its the simplest way to solve this problem.

Of course the right way to fix this is to change our horrible setpagedevice
code so that it works the way the spec says it should.

The fix described above resolves the memory warning for me as well.

No differences expected.