More work with setpagedevice
commit 55bd0a59caf26a59f852e2dec5b76ad399d9408b fixed the problems with
RedValues and friends, but at the cost of preventing these keys being set
by setpagedevice.

This commit allows them to be set by creating a 'write only' dictioanry
whose keys are removed from the initial device dictionary before we make it
read-only and send it to setpagedevice.

See the comments in the commit for further details.

No differences expected.