Address leaking font cache in PCL interpreter
Bug #695017 "pcfont_do_reset() leaks memory"

The PCL interpreter should free its font cache on exit, but it does not.
This is because of the fact that when we copy fonts the font cache of the
original font is copied to the font cache of the copied font. Later, when
a high level device wants to retrieve glyph information, the font code
uses the font cache from the copied font.

This doesn't immediately seem like a problem, it seems like we should be
able to defer freeing the font cache until after we've closed the device,
and that would satisfy pdfwrite. However, the way that PCL 'passthrough'
works, the PCL interpreter can be run 'inside' the PXL interpreter. If we
free the PCL interpreter font cache when it exits then pdfwrite has font
copies which are still pointing at the freed font cache.

We did discuss having the PCL/PXL interpreter set the gs_lib_ctx font_dir
(it currently dos not do so) and having pdfwrite set the font cache of the
fonts it copies use that, instead of the font cache from the original font.
But this seemed to me likely to run into the possibility of the same problem
recurring in future.

Instead, we have pdfwrite create and maintain its own font cache. When pdfwrite
copies a font it replaces the font cache pointer (which is initially the same
as the original font) with the new pdfwrite-specific font cache.

This means that any interpreter can free its font cache and pdfwrite will
still be able to use its own font cache when retrieving glyph information.

This commit includes both the changes to pdfwrite to enable the new font cache
and the changes to the PCL/PXL interpreter to free the font cache when the
interpreter is done with it.

No differences expected.